mod:移除nip相关问题

+++ scripts/install-traefik.sh
@@ -31,7 +31,6 @@ read -p "是否重装 traefik (y/n)" yN
 case $yN in
     [Yy]* )
     container_name=traefik
-    nip_enabled="false"
     TRAEFIK_AUTH_USER=$(`dirname $0`/get-args.sh TRAEFIK_AUTH_USER 用户名)
     if [ -z "$TRAEFIK_AUTH_USER" ]; then
         read -p "请输入用户名:" TRAEFIK_AUTH_USER
@@ -56,51 +55,56 @@ case $yN in
     echo "密码: $TRAEFIK_AUTH_PASSWORD"
     digest="$(printf "%s:%s:%s" "$TRAEFIK_AUTH_USER" "traefik" "$TRAEFIK_AUTH_PASSWORD" | md5sum | awk '{print $1}' )"
     userlist=$(printf "%s:%s:%s\n" "$TRAEFIK_AUTH_USER" "traefik" "$digest")
-    if [ "$tls" = "true" ]; then
-        # 检查domain结尾是否是nip.io
-        echo "检查domain结尾是否是nip.io"
-        if echo "$domain" | grep -q -E '\.nip.io$'
-        then
-            nip_enabled="true"
-            echo "使用nip.io域名"
-        else
-            echo "使用nip.io域名，跳过cf配置"
-            nip_enabled="false"
-            acme_email=$(`dirname $0`/get-args.sh acme_email acme的email)
-            if [ -z "$acme_email" ]; then
-                read -p "请输入acme的email: " acme_email
-                if [ -z "$acme_email" ]; then
-                    echo "acme的email不能为空"
-                    exit 1
-                fi
-                `dirname $0`/set-args.sh acme_email $acme_email
-            fi
-            CF_API_EMAIL=$(`dirname $0`/get-args.sh CF_API_EMAIL Cloudflare的邮箱)
-            if [ -z "$CF_API_EMAIL" ]; then
-                read -p "请输入Cloudflare的邮箱:" CF_API_EMAIL
-                if [ -z "$CF_API_EMAIL" ]; then
-                    echo "Cloudflare的邮箱不能为空"
-                    exit 1
-                fi
-                `dirname $0`/set-args.sh CF_API_EMAIL "$CF_API_EMAIL"
-            fi
-            CF_DNS_API_TOKEN=$(`dirname $0`/get-args.sh CF_DNS_API_TOKEN Cloudflare的api令牌)
-            if [ -z "$CF_DNS_API_TOKEN" ]; then
-                read -p "请输入Cloudflare的api令牌:" CF_DNS_API_TOKEN
-                if [ -z "$CF_DNS_API_TOKEN" ]; then
-                    echo "Cloudflare的api令牌不能为空"
-                    exit 1
-                fi
-                `dirname $0`/set-args.sh CF_DNS_API_TOKEN "$CF_DNS_API_TOKEN"
-            fi
+    acme_email=$(`dirname $0`/get-args.sh acme_email acme的email)
+    if [ -z "$acme_email" ]; then
+      read -p "请输入acme的email: " acme_email
+      if [ -z "$acme_email" ]; then
+          echo "acme的email不能为空"
+          exit 1
+      fi
+      `dirname $0`/set-args.sh acme_email $acme_email
+    fi
+    CF_API_EMAIL=$(`dirname $0`/get-args.sh CF_API_EMAIL Cloudflare的邮箱)
+    if [ -z "$CF_API_EMAIL" ]; then
+        read -p "请输入Cloudflare的邮箱:" CF_API_EMAIL
+        if [ -z "$CF_API_EMAIL" ]; then
+            echo "Cloudflare的邮箱不能为空"
+            exit 1
         fi
+        `dirname $0`/set-args.sh CF_API_EMAIL "$CF_API_EMAIL"
+    fi
+    CF_DNS_API_TOKEN=$(`dirname $0`/get-args.sh CF_DNS_API_TOKEN Cloudflare的api令牌)
+    if [ -z "$CF_DNS_API_TOKEN" ]; then
+        read -p "请输入Cloudflare的api令牌:" CF_DNS_API_TOKEN
+        if [ -z "$CF_DNS_API_TOKEN" ]; then
+            echo "Cloudflare的api令牌不能为空"
+            exit 1
+        fi
+        `dirname $0`/set-args.sh CF_DNS_API_TOKEN "$CF_DNS_API_TOKEN"
     fi
     echo "停止之前的traefik容器"
     container_name=traefik
     image=traefik
     docker pull ${image}
     docker ps -a -q --filter "name=$container_name" | grep -q . && docker rm -fv $container_name
-    echo "启动traefik容器" && \
+    echo "启动traefik容器"
+    # 使用cloudflare，需要设置CF_API_EMAIL和CF_DNS_API_TOKEN
+    CLOUDFLARE_ENVS="-e \"CF_API_EMAIL=${CF_API_EMAIL}\" -e \"CF_DNS_API_TOKEN=${CF_DNS_API_TOKEN}"
+    entrypoints_cmd="--entrypoints.web.address=\":80\""
+    if [ "$tls" = "true" ]; then
+      entrypoints_cmd="${entrypoints_cmd} --entrypoints.websecure.address=\":443\""
+      entrypoints_cmd="${entrypoints_cmd} --entrypoints.web.http.redirections.entryPoint.to=websecure"
+      entrypoints_cmd="${entrypoints_cmd} --entrypoints.web.http.redirections.entryPoint.scheme=https"
+    fi
+    certificatesresolvers_cmd=""
+    if [ "$tls" = "true" ]; then
+      certificatesresolvers_cmd="${certificatesresolvers_cmd} --certificatesresolvers.traefik.acme.dnschallenge=true"
+      certificatesresolvers_cmd="${certificatesresolvers_cmd} --certificatesresolvers.traefik.acme.dnschallenge.provider=cloudflare"
+      certificatesresolvers_cmd="${certificatesresolvers_cmd} --certificatesResolvers.traefik.acme.email=$acme_email"
+      certificatesresolvers_cmd="${certificatesresolvers_cmd} --certificatesresolvers.traefik.acme.storage=/acme/acme.json"
+    fi
+
+
     docker run --name=traefik \
     --restart=always -d -m 128M \
     -e TZ="Asia/Shanghai" \
@@ -109,23 +113,16 @@ case $yN in
     `if [ "$tls" = "true" ]; then echo  "-p 443:443/udp -p 443:443"; fi` \
     -e UID=`id -u` \
     -e GID=`id -g` \
-    `if [ "$tls" = "true" ]; then \
-        if [ "nip_enabeld" = "false" ];
-        then \
-            echo "-e \"CF_API_EMAIL=${CF_API_EMAIL}\" -e \"CF_DNS_API_TOKEN=${CF_DNS_API_TOKEN}\""; \
-        else \
-            echo ""; \
-        fi \
-    fi `\
-    --network=$docker_network_name --network-alias=traefik \
-    --label 'traefik.http.routers.traefik.rule=Host(`traefik'.$domain'`)' \
-    --label "traefik.http.routers.traefik.tls=${tls}" \
-    --label "traefik.http.routers.traefik.service=traefik" \
-    --label "traefik.http.routers.traefik.tls.certresolver=traefik" \
-    --label "traefik.http.routers.traefik.tls.domains[0].main=*.$domain" \
-    --label "traefik.http.services.traefik.loadbalancer.server.port=8080" \
-    --label "traefik.http.middlewares.traefik-auth.digestauth.users=$userlist" \
-    --label "traefik.http.routers.traefik.middlewares=traefik-auth@docker" \
+    ${CLOUDFLARE_ENVS} \
+    --network=$docker_network_name --hostname=${container_name} --network-alias=${container_name} \
+    --label 'traefik.http.routers.'${container_name}'.rule=Host(`'${container_name}.$domain'`)' \
+    --label "traefik.http.routers.${container_name}.tls=${tls}" \
+    --label "traefik.http.routers.${container_name}.service=${container_name}" \
+    --label "traefik.http.routers.${container_name}.tls.certresolver=traefik" \
+    --label "traefik.http.routers.${container_name}.tls.domains[0].main=*.$domain" \
+    --label "traefik.http.services.${container_name}.loadbalancer.server.port=8080" \
+    --label "traefik.http.middlewares.${container_name}-auth.digestauth.users=$userlist" \
+    --label "traefik.http.routers.${container_name}.middlewares=${container_name}-auth@docker" \
     --label "traefik.enable=true" \
     -v $base_data_dir/traefik/acme:/acme \
     -v $base_data_dir/traefik/config/providers:/config/providers \
@@ -138,30 +135,8 @@ case $yN in
     --providers.docker.endpoint=tcp://dockerproxy:2375 \
     --providers.docker.network=$docker_network_name \
     --providers.docker.exposedbydefault=false \
-    --entrypoints.web.address=":80" \
-    `if [ "$tls" = "true" ]; then \
-    echo """ \
-    --entrypoints.websecure.address=":443" \
-    --entrypoints.web.http.redirections.entryPoint.to=websecure \
-    --entrypoints.web.http.redirections.entryPoint.scheme=https \
-    """ \
-    fi` \
-    `if [ "$tls" = "true" ]; then \
-        if [ "nip_enabeld" = "false" ]; then \
-        echo """ \
-            --certificatesresolvers.traefik.acme.dnschallenge=true \
-            --certificatesresolvers.traefik.acme.dnschallenge.provider=cloudflare \
-            --certificatesResolvers.traefik.acme.dnsChallenge.delayBeforeCheck=10 \
-            --certificatesResolvers.traefik.acme.dnsChallenge.resolvers="1.1.1.1:53,8.8.8.8:53" \
-            --certificatesresolvers.traefik.acme.email=$acme_email \
-            --certificatesresolvers.traefik.acme.storage=/acme/acme.json \
-            """ \
-        else \
-            echo """ \
-            --certificatesresolvers.traefik=false \
-            """ \
-        fi \
-    fi` \
+    ${entrypoints_cmd} \
+    ${certificatesresolvers_cmd} \
     --providers.file.directory=/config/providers \
     --global.sendAnonymousUsage \
     --serverstransport.insecureskipverify=true \
